# documents/IndexUniverseService_OpenAPI.yaml
openapi: 3.0.3
info:
  title: OptiFrontier Index Universe Service API
  version: 1.0.0
  description: |
    API surface for managing and retrieving the top 10 S&P 500 universe snapshots.
servers:
  - url: https://api.optifrontier.example.com
    description: Production
  - url: https://staging.optifrontier.example.com
    description: Staging
security:
  - bearerAuth: []
tags:
  - name: Universe
    description: Operations that expose current and historical S&P 500 top 10 membership.
  - name: Jobs
    description: Operations that manage ingestion and validation workflows.
paths:
  /universe/current:
    get:
      tags:
        - Universe
      summary: Get current S&P 500 top 10 snapshot
      description: Returns the effective universe snapshot for the supplied tenant. Optionally supply an as-of date to retrieve a historical snapshot.
      parameters:
        - name: asOf
          in: query
          description: Trading date to evaluate (ISO 8601 date). Defaults to most recent effective snapshot.
          required: false
          schema:
            type: string
            format: date
        - name: includeConstituentMeta
          in: query
          description: When true, response includes identifier metadata (CUSIP, ISIN).
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Snapshot located.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UniverseSnapshot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /universe/history:
    get:
      tags:
        - Universe
      summary: List S&P 500 top 10 snapshots over a date range
      description: Returns paginated universe snapshots for the requested date window ordered by as_of_date descending.
      parameters:
        - name: startDate
          in: query
          description: Inclusive start date (ISO 8601 date). Defaults to 30 days before endDate.
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Inclusive end date (ISO 8601 date). Defaults to current date.
          required: false
          schema:
            type: string
            format: date
        - name: limit
          in: query
          description: Page size (1-100).
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: cursor
          in: query
          description: Pagination cursor from a previous response.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Page of snapshots.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UniverseSnapshotPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /universe/rebalance-job:
    post:
      tags:
        - Jobs
      summary: Trigger a manual rebalance ingestion job
      description: Enqueues a manual or override ingestion workflow. Requires portfolio.admin role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RebalanceJobRequest'
      responses:
        '202':
          description: Job accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebalanceJobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UniverseSnapshot:
      type: object
      required:
        - snapshotId
        - tenantId
        - asOfDate
        - effectiveAt
        - publishedAt
        - source
        - constituents
      properties:
        snapshotId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        asOfDate:
          type: string
          format: date
        effectiveAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
        source:
          type: string
          enum:
            - spdj
            - manual_override
        hash:
          type: string
          description: SHA256 hash of ordered constituents payload.
        constituents:
          type: array
          minItems: 1
          maxItems: 10
          items:
            $ref: '#/components/schemas/Constituent'
    Constituent:
      type: object
      required:
        - position
        - ticker
        - weight
        - freeFloatMarketCap
      properties:
        position:
          type: integer
          minimum: 1
          maximum: 10
        ticker:
          type: string
          pattern: '^[A-Z.]{1,10}$'
        cusip:
          type: string
          pattern: '^[0-9A-Z]{9}$'
        isin:
          type: string
          pattern: '^[A-Z]{2}[A-Z0-9]{9}[0-9]$'
        weight:
          type: number
          format: double
          minimum: 0
          maximum: 1
        freeFloatMarketCap:
          type: number
          format: double
          minimum: 0
        sector:
          type: string
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
    UniverseSnapshotPage:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UniverseSnapshot'
        nextCursor:
          type: string
          nullable: true
        totalCount:
          type: integer
          nullable: true
    RebalanceJobRequest:
      type: object
      properties:
        effectiveDate:
          type: string
          format: date
          description: Override effective date to backfill historical corrections.
        source:
          type: string
          enum:
            - manual_override
            - provider_replay
          default: manual_override
        force:
          type: boolean
          description: Skip duplicate detection and force job enqueue.
          default: false
      additionalProperties: false
    RebalanceJobResponse:
      type: object
      required:
        - jobId
        - status
        - requestedAt
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - queued
            - running
            - succeeded
            - failed
        requestedAt:
          type: string
          format: date-time
        effectiveDate:
          type: string
          format: date
          nullable: true
        queueEventId:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Caller lacks permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflicting job already exists.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Throttled.
      headers:
        Retry-After:
          schema:
            type: string
          description: Backoff hint in seconds or HTTP date.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'